Auth project with components
Auth app will not load as there is an issue in publicPath in webpack. we will fix it next commits

Note: In Marketing App it worked without public path because of (ref. image -51) if publicPath is not mentioned it's going to fetch default by relative to remoteEntry.js. So as Marketing doesn't have nested routes it worked perfectly.

But in Auth app we have nested routes
so (ref image -49) it tries to fetch main.js by adding main.js at the end of route which won't work.
ex: https://localhost:8082/auth/sigin/main.js

solution: (image - 50)